% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.r
\name{links}
\alias{links}
\title{Creates HTML links to an UCSC Genome Browser from a feature file.}
\usage{
links(i, base = NULL, org = NULL, db = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf> > out.html}

\item{base}{The browser basename.  Default: http://genome.ucsc.edu}

\item{org}{The organism. Default: human}

\item{db}{The build.  Default: hg18}
}
\description{
Creates HTML links to an UCSC Genome Browser from a feature file.
}
